requirements.txt
requests
pandas
matplotlib
openai
python-dotenv